@@ -1,15 +1,16 @@
 name: receive_sharing_intent
 description: A flutter plugin that enables flutter apps to receive sharing photos, text or url from other apps.
-version: 1.4.5
+version: 1.8.0
 homepage: https://github.com/KasemJaffer/receive_sharing_intent
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <4.0.0"
   flutter: ">=1.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  plugin_platform_interface: ^2.1.8
 
 dev_dependencies:
   flutter_test:
