@@ -7,6 +7,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:receive_sharing_intent/receive_sharing_intent.dart';
 
 import 'package:receive_sharing_intent_example/main.dart';
 
@@ -15,11 +16,17 @@ void main() {
     // Build our app and trigger a frame.
     await tester.pumpWidget(MyApp());
 
+    // Set mock values for ReceiveSharingIntent
+    ReceiveSharingIntent.setMockValues(
+      initialMedia: [],
+      mediaStream: Stream.empty(),
+    );
+
     // Verify that platform version is retrieved.
     expect(
       find.byWidgetPredicate(
-        (Widget widget) => widget is Text &&
-                           widget.data?.startsWith('Running on:') == true,
+        (Widget widget) =>
+            widget is Text && widget.data?.startsWith('Running on:') == true,
       ),
       findsOneWidget,
     );
