@@ -1,18 +1,31 @@
+library receive_sharing_intent;
+
 import 'dart:async';
-import 'dart:convert';
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
+import 'package:plugin_platform_interface/plugin_platform_interface.dart';
+import 'src/receive_sharing_intent_mobile.dart';
 
-class ReceiveSharingIntent {
-  static const MethodChannel _mChannel =
-  const MethodChannel('receive_sharing_intent/messages');
-  static const EventChannel _eChannelMedia =
-  const EventChannel("receive_sharing_intent/events-media");
-  static const EventChannel _eChannelLink =
-  const EventChannel("receive_sharing_intent/events-text");
+part 'src/data/shared_media_file.dart';
 
-  static Stream<List<SharedMediaFile>>? _streamMedia;
-  static Stream<String>? _streamLink;
+abstract class ReceiveSharingIntent extends PlatformInterface {
+  ReceiveSharingIntent() : super(token: _token);
+
+  static final Object _token = Object();
+
+  static ReceiveSharingIntent _instance = ReceiveSharingIntentMobile();
+
+  /// The default instance of [ReceiveSharingIntent] to use.
+  static ReceiveSharingIntent get instance => _instance;
+
+  /// Platform-specific implementations should set this to their own
+  /// platform-specific class that extends [SamplePluginPlatform] when they
+  /// register themselves.
+  static set instance(ReceiveSharingIntent instance) {
+    PlatformInterface.verify(instance, _token);
+    _instance = instance;
+  }
 
   /// Returns a [Future], which completes to one of the following:
   ///
@@ -21,32 +34,8 @@ class ReceiveSharingIntent {
   ///
   /// NOTE. The returned media on iOS (iOS ONLY) is already copied to a temp folder.
   /// So, you need to delete the file after you finish using it
-  static Future<List<SharedMediaFile>> getInitialMedia() async {
-    final json = await _mChannel.invokeMethod('getInitialMedia');
-    if (json == null) return [];
-    final encoded = jsonDecode(json);
-    return encoded
-        .map<SharedMediaFile>((file) => SharedMediaFile.fromJson(file))
-        .toList();
-  }
-
-  /// Returns a [Future], which completes to one of the following:
-  ///
-  ///   * the initially stored link (possibly null), on successful invocation;
-  ///   * a [PlatformException], if the invocation failed in the platform plugin.
-  static Future<String?> getInitialText() async {
-    return await _mChannel.invokeMethod('getInitialText');
-  }
-
-  /// A convenience method that returns the initially stored link
-  /// as a new [Uri] object.
-  ///
-  /// If the link is not valid as a URI or URI reference,
-  /// a [FormatException] is thrown.
-  static Future<Uri?> getInitialTextAsUri() async {
-    final data = await getInitialText();
-    if (data == null) return null;
-    return Uri.parse(data);
+  Future<List<SharedMediaFile>> getInitialMedia() {
+    throw UnimplementedError('getInitialMedia() has not been implemented.');
   }
 
   /// Sets up a broadcast stream for receiving incoming media share change events.
@@ -65,99 +54,51 @@ class ReceiveSharingIntent {
   ///
   /// If the app was started by a link intent or user activity the stream will
   /// not emit that initial one - query either the `getInitialMedia` instead.
-  static Stream<List<SharedMediaFile>> getMediaStream() {
-    if (_streamMedia == null) {
-      final stream =
-      _eChannelMedia.receiveBroadcastStream("media").cast<String?>();
-      _streamMedia = stream.transform<List<SharedMediaFile>>(
-        new StreamTransformer<String?, List<SharedMediaFile>>.fromHandlers(
-          handleData: (String? data, EventSink<List<SharedMediaFile>> sink) {
-            if (data == null) {
-              sink.add([]);
-            } else {
-              final encoded = jsonDecode(data);
-              sink.add(encoded
-                  .map<SharedMediaFile>(
-                      (file) => SharedMediaFile.fromJson(file))
-                  .toList());
-            }
-          },
-        ),
-      );
-    }
-    return _streamMedia!;
-  }
-
-  /// Sets up a broadcast stream for receiving incoming link change events.
-  ///
-  /// Returns a broadcast [Stream] which emits events to listeners as follows:
-  ///
-  ///   * a decoded data ([String]) event (possibly null) for each successful
-  ///   event received from the platform plugin;
-  ///   * an error event containing a [PlatformException] for each error event
-  ///   received from the platform plugin.
-  ///
-  /// Errors occurring during stream activation or deactivation are reported
-  /// through the `FlutterError` facility. Stream activation happens only when
-  /// stream listener count changes from 0 to 1. Stream deactivation happens
-  /// only when stream listener count changes from 1 to 0.
-  ///
-  /// If the app was started by a link intent or user activity the stream will
-  /// not emit that initial one - query either the `getInitialText` instead.
-  static Stream<String> getTextStream() {
-    if (_streamLink == null) {
-      _streamLink = _eChannelLink.receiveBroadcastStream("text").cast<String>();
-    }
-    return _streamLink!;
-  }
-
-  /// A convenience transformation of the stream to a `Stream<Uri>`.
-  ///
-  /// If the value is not valid as a URI or URI reference,
-  /// a [FormatException] is thrown.
-  ///
-  /// Refer to `getTextStream` about error/exception details.
-  ///
-  /// If the app was started by a share intent or user activity the stream will
-  /// not emit that initial uri - query either the `getInitialTextAsUri` instead.
-  static Stream<Uri> getTextStreamAsUri() {
-    return getTextStream().transform<Uri>(
-      new StreamTransformer<String, Uri>.fromHandlers(
-        handleData: (String data, EventSink<Uri> sink) {
-          sink.add(Uri.parse(data));
-        },
-      ),
-    );
+  Stream<List<SharedMediaFile>> getMediaStream() {
+    throw UnimplementedError('getMediaStream() has not been implemented.');
   }
 
   /// Call this method if you already consumed the callback
   /// and don't want the same callback again
-  static void reset() {
-    _mChannel.invokeMethod('reset').then((_) {});
+  Future<dynamic> reset() {
+    throw UnimplementedError('reset() has not been implemented.');
+  }
+
+  /// Initializes the plugin and sets the mock values for testing.
+  @visibleForTesting
+  static void setMockValues({
+    required List<SharedMediaFile> initialMedia,
+    required Stream<List<SharedMediaFile>> mediaStream,
+  }) {
+    ReceiveSharingIntent.instance = _ReceiveSharingIntentMock(
+      initialMedia: List.from(initialMedia),
+      mediaStream: mediaStream,
+    );
   }
 }
 
-class SharedMediaFile {
-  /// Image or Video path.
-  /// NOTE. for iOS only the file is always copied
-  final String path;
+/// A mock implementation of [ReceiveSharingIntent] for testing.
+class _ReceiveSharingIntentMock extends ReceiveSharingIntent {
+  final List<SharedMediaFile> initialMedia;
+  final Stream<List<SharedMediaFile>> mediaStream;
 
-  /// Video thumbnail
-  final String? thumbnail;
+  _ReceiveSharingIntentMock({
+    required this.initialMedia,
+    required this.mediaStream,
+  });
 
-  /// Video duration in milliseconds
-  final int? duration;
+  @override
+  Future<List<SharedMediaFile>> getInitialMedia() async {
+    return initialMedia;
+  }
 
-  /// Whether its a video or image or file
-  final SharedMediaType type;
+  @override
+  Stream<List<SharedMediaFile>> getMediaStream() {
+    return mediaStream;
+  }
 
-  SharedMediaFile(this.path, this.thumbnail, this.duration, this.type);
-
-  SharedMediaFile.fromJson(Map<String, dynamic> json)
-      : path = json['path'],
-        thumbnail = json['thumbnail'],
-        duration = json['duration'],
-        type = SharedMediaType.values[json['type']];
+  @override
+  Future<dynamic> reset() async {
+    return initialMedia.clear();
+  }
 }
-
-enum SharedMediaType { IMAGE, VIDEO, FILE }
